Adapt other GCs to central memory limit functions

+++ rts/motoko-rts/src/gc/incremental/write_barrier.rs
@@ -17,7 +17,7 @@ pub unsafe fn pre_write_barrier<M: Memory>(mem: &mut M, location: *mut Value) {
     debug_assert!(!is_skewed(location as u32));
     debug_assert_ne!(location, core::ptr::null_mut());
     let value = *location;
-    if value.is_ptr() && value.get_ptr() >= mem.heap_base() {
+    if value.is_ptr() && value.get_ptr() >= mem.heap_base() as usize {
         IncrementalGC::instance(mem).pre_write_barrier(value);
     }
 }
